Refactored project structure (everything API related), added explicit Task-Class which is parsed when a Task is queried, added RestRequestListNew and RestRequestListModify for operations on lists
